Enhance receipt processing, management, and UI components

+++ Backend/src/ProzoroBanka.Application/Campaigns/Commands/UpdateCampaign/UpdateCampaignCommand.cs
@@ -12,7 +12,7 @@ public record UpdateCampaignCommand(
 	Guid CampaignId,
 	string? Title,
 	string? Description,
-	decimal? GoalAmount,
+	long? GoalAmount,
 	DateTime? Deadline,
 	string? SendUrl) : IRequest<ServiceResponse<CampaignDto>>, ICacheInvalidatingCommand
 {
